Bit manipulation tasks

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -11,10 +11,11 @@
 
 int set_bit(unsigned long int *n, unsigned int index)
 {
-	unsigned long int *p;
+	unsigned int *p;
 
 	if (index > 64)
 		return (-1);
+
 	p = (n >> index) & 1;
 	return (1);
 }
